requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
setuptools==69.0.0